viettx day code fix bug
- Thay doi can phan phoi  chuyen sang thu tuc
- fix bug them moi phieu bom
- validate ngay bat dau va ngay ket thuc trong 1 thang

+++ CKPManagement/src/main/java/com/dvd/ckp/business/dao/RentEquimentDAO.java
@@ -1,5 +1,6 @@
 package com.dvd.ckp.business.dao;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import com.dvd.ckp.domain.Rent;
@@ -16,5 +17,5 @@ public interface RentEquimentDAO {
 	
 	public List<Rent> storeRent(Long rentID);
 	
-	public Long getMaxID();
+	public BigInteger getMaxID();
 }
